Replace `exec` by `run`
With the `exec` interpretation, the I/O computation of a process would
only finish when execution ends. To ensure this behavior in the presence
of parallel composition, `MVar`s had to be created for synchronization
purposes. When processes were running indefinitely, this could lead to
certain `MVar`s never being released.

With the `run` interpretation, the I/O computation of a process finishes
immediately. As a result, no synchronization through `MVar`s is needed,
and the above-mentioned problem does not arise.