Port `Communication` to the Þ-calculus
On the way, we also introduce a few improvements.

We replace the proofs of most of the steps in the proof of
`duploss_detour_collapse` by `sorry`, since we cannot prove these steps
with the current proof methodes and crafting manual proofs would be too
labor-intensive. We shall fill these newly introduced holes once we have
better support with proof methods.